+++ source/dotnet/wholesale-api/Batches/Batches.UnitTests/Batches.UnitTests.csproj
@@ -16,6 +16,7 @@
       <Folder Include="Infrastructure" />
     </ItemGroup>
     <ItemGroup>
+      <ProjectReference Include="..\..\IntegrationEventPublishing\IntegrationEventPublishing.Infrastructure\IntegrationEventPublishing.Infrastructure.csproj" />
       <ProjectReference Include="..\..\Test.Core\Test.Core.csproj" />
       <ProjectReference Include="..\Batches.Application\Batches.Application.csproj" />
       <ProjectReference Include="..\Batches.Infrastructure\Batches.Infrastructure.csproj" />
