--- source/dotnet/wholesale-api/IntegrationEventPublishing/IntegrationEventsPublising.Application.Tests/IntegrationEventsPublising.Application.Tests.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-    <PropertyGroup>
-      <AssemblyName>Energinet.DataHub.Wholesale.IntegrationEventsPublising.Application.Tests</AssemblyName>
-      <RootNamespace>Energinet.DataHub.Wholesale.IntegrationEventsPublising.Application.Tests</RootNamespace>
-        <IsPackable>false</IsPackable>
-        <IsTestProject>true</IsTestProject>
-    </PropertyGroup>
-    <ItemGroup>
-        <PackageReference Include="Energinet.DataHub.Core.TestCommon" Version="4.1.0" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.5.0" />
-        <PackageReference Include="NUnit" Version="3.13.3" />
-        <PackageReference Include="NUnit3TestAdapter" Version="4.4.2" />
-        <PackageReference Include="NUnit.Analyzers" Version="3.6.1" />
-        <PackageReference Include="coverlet.collector" Version="3.2.0" />
-    </ItemGroup>
-</Project>
