+++ source/dotnet/wholesale-api/IntegrationEventPublishing/IntegrationEventsPublising.IntegrationTests/IntegrationEventsPublising.IntegrationTests.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+      <AssemblyName>Energinet.DataHub.Wholesale.IntegrationEventsPublishing.IntegrationTests</AssemblyName>
+      <RootNamespace>Energinet.DataHub.Wholesale.IntegrationEventsPublishing.IntegrationTests</RootNamespace>
+    </PropertyGroup>
+    <ItemGroup>
+        <PackageReference Include="Energinet.DataHub.Core.TestCommon" Version="4.1.0" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.5.0" />
+        <PackageReference Include="NUnit" Version="3.13.3" />
+        <PackageReference Include="NUnit3TestAdapter" Version="4.4.2" />
+        <PackageReference Include="NUnit.Analyzers" Version="3.6.1" />
+        <PackageReference Include="coverlet.collector" Version="3.2.0" />
+    </ItemGroup>
+</Project>
