+++ source/dotnet/wholesale-api/IntegrationEventPublishing/IntegrationEventsPublising.UnitTests/IntegrationEventsPublising.UnitTests.csproj
@@ -1,0 +1,29 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+      <AssemblyName>Energinet.DataHub.Wholesale.IntegrationEventsPublishing.UnitTests</AssemblyName>
+      <RootNamespace>Energinet.DataHub.Wholesale.IntegrationEventsPublishing.UnitTests</RootNamespace>
+        <IsPackable>false</IsPackable>
+        <IsTestProject>true</IsTestProject>
+    </PropertyGroup>
+    <ItemGroup>
+        <PackageReference Include="Energinet.DataHub.Core.TestCommon" Version="4.1.0" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.5.0" />
+        <PackageReference Include="NUnit" Version="3.13.3" />
+        <PackageReference Include="NUnit3TestAdapter" Version="4.4.2" />
+        <PackageReference Include="NUnit.Analyzers" Version="3.6.1" />
+        <PackageReference Include="coverlet.collector" Version="3.2.0" />
+    </ItemGroup>
+    <ItemGroup>
+      <EmbeddedResource Include="..\..\..\..\databricks\calculation-engine\contracts\calculator-result.json">
+        <Link>Integration\CalculationOutput\calculator-result.json</Link>
+      </EmbeddedResource>
+      <EmbeddedResource Include="..\..\..\..\databricks\calculation-engine\contracts\time-series-point-quality.json">
+        <Link>Integration\CalculationOutput\time-series-point-quality.json</Link>
+      </EmbeddedResource>
+    </ItemGroup>
+    <ItemGroup>
+      <ProjectReference Include="..\..\Test.Core\Test.Core.csproj" />
+      <ProjectReference Include="..\..\WebApi\WebApi.csproj" />
+      <ProjectReference Include="..\IntegrationEventPublishing.Infrastructure\IntegrationEventPublishing.Infrastructure.csproj" />
+    </ItemGroup>
+</Project>
