--- source/dotnet/wholesale-api/WebApi.UnitTests/Domain/ProcessAggregate/ProcessTests.cs
@@ -1,33 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
-using FluentAssertions;
-using Xunit;
-
-namespace Energinet.DataHub.Wholesale.WebApi.UnitTests.Domain.ProcessAggregate;
-
-public class ProcessTests
-{
-    [Fact]
-    public void MustEnsureThatProcessTypeCompliesWithTheContract()
-    {
-        // Arrange & Act
-        var domainProcessTypeValues = Enum.GetValues(typeof(ProcessType));
-        var contractProcessTypeValues = Enum.GetValues(typeof(Batches.Interfaces.Models.ProcessType));
-
-        // Assert
-        domainProcessTypeValues.Should().BeEquivalentTo(contractProcessTypeValues);
-    }
-}
